pyproject.toml
[tool.poetry]
name = "gpustackhelper"
version = "0.1.0"
description = "The GPUStack Installer is a helper tool for installing and configuring GPUStack."
authors = ["GPUStack Authors <contact@gpustack.ai>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
gpustack = { path = '/Users/dengyuxing/git/github.com/gpustack/gpustack/dist/gpustack-0.0.0-py3-none-any.whl' }
# gpustack = { version = "0.6.1", extras = ["audio"] }

PySide6-Essentials="~6.8"
platformdirs = "^4.3.8"

[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
helper = "gpustackhelper.main:main"
